requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiounittest==1.5.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
fake-useragent==2.2.0
frozenlist==1.8.0
howlongtobeatpy==1.0.20
idna==3.11
multidict==6.7.1
numpy==2.4.2
pandas==3.0.1
pip==25.3
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wrapt==2.1.2
yarl==1.23.0
